refactor(debug): keep debug-session CONTRACT in types, move fs IO to consumers
@lmnr-ai/types is type-only again: debug-session.ts now exports just the
DebugSessionFile interface + DEBUG_SESSION_DIR/DEBUG_SESSION_FILE consts. The
sync read/write helpers move to each runtime consumer — SDK
src/debug/debug-session-file.ts (read+write, with a leaner str() coercion read)
and CLI src/utils/debug-session-file.ts (write). Both import the shared
contract, so the on-disk shape can't drift between writers.

Removes the @types/node devDep + types:[node] tsconfig hack that the in-types
node:fs helpers had forced into a type-only package; types now passes tsc
--noEmit with no node lib.
