+++ packages/types/package.json
@@ -36,8 +36,7 @@
   "author": "founders@lmnr.ai",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@opentelemetry/sdk-trace-base": "^1.30.1 || ^2.0.0",
-    "@types/node": "^24.12.0"
+    "@opentelemetry/sdk-trace-base": "^1.30.1 || ^2.0.0"
   },
   "peerDependencies": {
     "@opentelemetry/sdk-trace-base": "^1.30.1 || ^2.0.0"
