+++ packages/types/tsconfig.json
@@ -2,10 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist",
-    "rootDir": "src",
-    "types": [
-      "node"
-    ]
+    "rootDir": "src"
   },
   "include": [
     "src/**/*"
